bug fix for right offscreen nav
on the right navigation demo, in the latest firefox + chrome, a horizontal scrollbar appeared on the page that allowed for scrolling to the right to see the hidden navigation at any time on all screen sizes.

This bug also appeared on iOS Safari, where at any point a user could swipe right to reveal the hidden navigation.

adding a body { overflow-x: hidden; } fixes this issue and all other demos continue to work without issue.